pyproject.toml
[project]
name = "hermes-openclaw-orchestrator"
version = "0.1.0"
description = "Hermes-as-Master / OpenClaw-SubAgent Orchestration Layer"
authors = [{name = "Your Name"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "redis>=5.2.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "sse-starlette>=2.2.0",
    "structlog>=24.0.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "pytest-asyncio>=0.24.0", "ruff>=0.8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator", "agents", "api", "shared"]

[tool.ruff]
line-length = 100
target-version = "py311"
